chore: rename ext-airflow -> airflow-ext
