+++ airflow_ext/main.py
@@ -8,7 +8,7 @@ import typer
 from meltano_extension_sdk.extension import DescribeFormat
 from meltano_extension_sdk.logging import default_logging_config, parse_log_level
 
-from ext_airflow.wrapper import Airflow
+from airflow_ext.wrapper import Airflow
 
 log = structlog.get_logger("airflow_extension")
