+++ cookiecutter/wrapper-template/cookiecutter.json
@@ -2,7 +2,7 @@
   "source_name": "MyExtensionName",
   "admin_name": "FirstName LastName",
   "extension_name": "{{ cookiecutter.source_name.lower() }}",
-  "wrapper_id": "ext-{{ cookiecutter.extension_name }}",
+  "wrapper_id": "{{ cookiecutter.extension_name }}-ext",
   "library_name": "{{ cookiecutter.wrapper_id.replace('-', '_') }}",
   "cli_prefix": "{{ cookiecutter.extension_name.lower() }}",
   "wrapper_target_name": "some-third-party-cli"
