+++ cookiecutter/wrapper-template/{{cookiecutter.wrapper_id}}/pyproject.toml
@@ -12,7 +12,7 @@ PyYAML = "^6.0.0"
 pydantic = "^1.9.0"
 click = "^8.1.3"
 typer = "^0.6.1"
-ext-airflow = {git = "https://github.com/meltano/ext-airflow.git", rev = "feat-init-extension"} # TODO: update before merge
+airflow-ext = {git = "https://github.com/meltano/airflow-ext.git"}
 
 [tool.poetry.dev-dependencies]
 black = "^22.3.0"
